Version updated 1.0.1 -- batch operations added for optimization purpose

+++ src/jStorageReact.ts
@@ -1,31 +1,34 @@
-// jStorageReact.ts
-
 interface JStorageMeta {
   CRC32: { [key: string]: number };
   TTL: { [key: string]: number };
 }
 
 class JStorageReact {
-  version: string;
-  storage: Storage | null;
-  jStorage: { [key: string]: any };
-  jStorageMeta: JStorageMeta;
-  crc32Table: number[];
+  version: string = "1.0.0";
+  storage: Storage | null = this.isBrowser() ? window.localStorage : null;
+  jStorage: { [key: string]: any } = this.getStoredJStorage();
+  jStorageMeta: JStorageMeta = this.jStorage.__jstorage_meta || {
+    CRC32: {},
+    TTL: {},
+  };
+  crc32Table: number[] = [];
+  batchChanges: boolean = false;
+  batchQueue: {
+    type: string;
+    key: string;
+    value?: any;
+    options?: { TTL?: number };
+  }[] = [];
 
   constructor() {
-    this.version = "1.0.0";
-    this.storage = this.isBrowser() ? window.localStorage : null;
-    this.jStorage = this.storage ? this.getStoredJStorage() : {};
-    this.jStorageMeta = this.jStorage.__jstorage_meta || { CRC32: {}, TTL: {} };
-    this.crc32Table = [];
     this.init();
   }
 
-  private isBrowser() {
+  private isBrowser(): boolean {
     return typeof window !== "undefined";
   }
 
-  private getStoredJStorage() {
+  private getStoredJStorage(): { [key: string]: any } {
     try {
       const storedData = this.storage?.getItem("jStorage");
       return storedData
@@ -45,14 +48,43 @@ class JStorageReact {
     }
   }
 
-  private save() {
-    if (this.storage) {
-      try {
-        this.storage.setItem("jStorage", JSON.stringify(this.jStorage));
-      } catch (error) {
-        console.error("Error saving jStorage data:", error);
-      }
+  private save(): void {
+    if (!this.storage) {
+      console.warn("Cannot save jStorage: localStorage is not available.");
+      return;
     }
+    try {
+      this.storage.setItem("jStorage", JSON.stringify(this.jStorage));
+    } catch (error) {
+      console.error("Error saving jStorage data:", error);
+    }
+  }
+
+  private processBatchQueue() {
+    this.batchQueue.forEach((item) => {
+      switch (item.type) {
+        case "set":
+          this.set(item.key, item.value, item.options);
+          break;
+        case "delete":
+          this.deleteKey(item.key);
+          break;
+        case "setTTL":
+          this.setTTL(item.key, item.options?.TTL || 0);
+          break;
+      }
+    });
+    this.batchQueue = [];
+    this.save();
+  }
+
+  startBatch(): void {
+    this.batchChanges = true;
+  }
+
+  endBatch(): void {
+    this.batchChanges = false;
+    this.processBatchQueue();
   }
 
   set(key: string, value: any, options: { TTL?: number } = {}): void {
@@ -60,29 +92,31 @@ class JStorageReact {
       console.warn("Cannot set jStorage value: localStorage is not available.");
       return;
     }
-
+    if (this.batchChanges) {
+      this.batchQueue.push({ type: "set", key, value, options });
+      return;
+    }
     if (typeof value === "undefined") {
       this.deleteKey(key);
       return;
     }
-
     if (typeof value === "function") {
       console.warn(
         "Cannot set jStorage value: function types are not supported."
       );
       return;
     }
-
     if (this.isXML(value)) {
       value = { _is_xml: true, xml: this.encodeXML(value) };
     } else if (typeof value === "object") {
       value = JSON.parse(JSON.stringify(value));
     }
-
     this.jStorage[key] = value;
     this.jStorageMeta.CRC32[key] = this.crc32(JSON.stringify(value));
     this.setTTL(key, options.TTL || 0);
-    this.save();
+    if (!this.batchChanges) {
+      this.save();
+    }
   }
 
   get(key: string, defaultValue: any = null): any {
@@ -90,7 +124,6 @@ class JStorageReact {
       console.warn("Cannot get jStorage value: localStorage is not available.");
       return defaultValue;
     }
-
     if (this.jStorage[key] && this.jStorage[key]._is_xml) {
       return this.decodeXML(this.jStorage[key].xml);
     }
@@ -104,11 +137,16 @@ class JStorageReact {
       );
       return;
     }
-
+    if (this.batchChanges) {
+      this.batchQueue.push({ type: "delete", key });
+      return;
+    }
     delete this.jStorage[key];
     delete this.jStorageMeta.TTL[key];
     delete this.jStorageMeta.CRC32[key];
-    this.save();
+    if (!this.batchChanges) {
+      this.save();
+    }
   }
 
   setTTL(key: string, ttl: number): void {
@@ -118,14 +156,19 @@ class JStorageReact {
       );
       return;
     }
-
+    if (this.batchChanges) {
+      this.batchQueue.push({ type: "setTTL", key, options: { TTL: ttl } });
+      return;
+    }
     if (!ttl) {
       delete this.jStorageMeta.TTL[key];
     } else {
       this.jStorageMeta.TTL[key] = Date.now() + ttl;
     }
     this.cleanupTTL();
-    this.save();
+    if (!this.batchChanges) {
+      this.save();
+    }
   }
 
   getTTL(key: string): number {
@@ -135,7 +178,6 @@ class JStorageReact {
       );
       return 0;
     }
-
     return this.jStorageMeta.TTL[key] - Date.now();
   }
 
@@ -143,7 +185,6 @@ class JStorageReact {
     if (!this.storage) {
       return;
     }
-
     const now = Date.now();
     for (const key in this.jStorageMeta.TTL) {
       if (this.jStorageMeta.TTL[key] <= now) {
@@ -174,13 +215,11 @@ class JStorageReact {
       j;
     const table =
       this.crc32Table || (this.crc32Table = this.generateCRC32Table());
-
     crc = crc ^ -1;
     for (i = 0; i < str.length; i++) {
       j = (crc ^ str.charCodeAt(i)) & 0xff;
       crc = (crc >>> 8) ^ table[j];
     }
-
     return (crc ^ -1) >>> 0;
   }
 
@@ -206,7 +245,6 @@ class JStorageReact {
       console.warn("Cannot get storage size: localStorage is not available.");
       return 0;
     }
-
     return new Blob([JSON.stringify(this.jStorage)]).size;
   }
 
@@ -219,7 +257,6 @@ class JStorageReact {
       console.warn("Cannot flush jStorage: localStorage is not available.");
       return;
     }
-
     this.jStorage = { __jstorage_meta: { CRC32: {}, TTL: {} } };
     this.save();
   }
